requirements_dev.txt
Django==4.1.3
invoke==1.7.3
PyYAML==6.0
requests==2.28.1
selenium==4.6.0
